+++ mycodo/mycodo_flask/routes_page.py
@@ -86,6 +86,7 @@ from mycodo.mycodo_flask.utils import utils_math
 from mycodo.mycodo_flask.utils import utils_output
 from mycodo.mycodo_flask.utils import utils_pid
 from mycodo.mycodo_flask.utils import utils_timer
+from mycodo.utils.sunriseset import Sun
 from mycodo.utils.system_pi import add_custom_measurements
 from mycodo.utils.system_pi import csv_to_list_of_int
 from mycodo.utils.system_pi import list_to_csv
@@ -887,6 +888,35 @@ def page_function():
     form_conditional = forms_conditional.Conditional()
     form_conditional_actions = forms_conditional.ConditionalActions()
 
+    # Calculate sunrise/sunset times if conditional controller is set up properly
+    sunrise_sunset_calculated = {}
+    for each_conditional in conditional:
+        if each_conditional.conditional_type == 'conditional_sunrise_sunset':
+            sunrise_sunset_calculated[each_conditional.id] = {}
+            try:
+                # Adjust for date offset
+                now = datetime.datetime.now()
+                new_date = now + datetime.timedelta(days=each_conditional.date_offset_days)
+
+                sun = Sun(latitude=each_conditional.latitude,
+                          longitude=each_conditional.longitude,
+                          zenith=each_conditional.zenith,
+                          day=new_date.day,
+                          month=new_date.month,
+                          year=new_date.year)
+                sunrise = sun.get_sunrise_time()
+                sunset = sun.get_sunset_time()
+
+                # Adjust for time offset
+                new_sunrise = sunrise['time_local'] + datetime.timedelta(minutes=each_conditional.time_offset_minutes)
+                new_sunset = sunset['time_local'] + datetime.timedelta(minutes=each_conditional.time_offset_minutes)
+
+                sunrise_sunset_calculated[each_conditional.id]['rise'] = 'Sunrise: {rise}'.format(rise=new_sunrise.strftime("%Y-%m-%d %H:%M"))
+                sunrise_sunset_calculated[each_conditional.id]['set'] = 'Sunset: {set}'.format(set=new_sunset.strftime("%Y-%m-%d %H:%M"))
+            except:
+                sunrise_sunset_calculated[each_conditional.id]['rise'] = None
+                sunrise_sunset_calculated[each_conditional.id]['set'] = None
+
     if request.method == 'POST':
         if not utils_general.user_has_permission('edit_controllers'):
             return redirect(url_for('routes_general.home'))
@@ -987,7 +1017,8 @@ def page_function():
                            output=output,
                            pid=pid,
                            units=MEASUREMENT_UNITS,
-                           user=user)
+                           user=user,
+                           sunrise_sunset_calculated=sunrise_sunset_calculated)
 
 
 @blueprint.route('/output', methods=('GET', 'POST'))