+++ mycodo/mycodo_flask/utils/utils_function.py
@@ -39,7 +39,8 @@ def func_add(form_add_func):
                     db.session.commit()
             elif form_add_func.func_type.data in ['conditional_measurement',
                                                   'conditional_output',
-                                                  'conditional_edge']:
+                                                  'conditional_edge',
+                                                  'conditional_sunrise_sunset']:
                 new_func = Conditional()
                 new_func.conditional_type = form_add_func.func_type.data
                 new_func.save()