+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
  - Add ability to convert Input measurements between units ([#346](https://github.com/kizniche/mycodo/issues/346))
  - Add ability to select whether lowering PID outputs are stored as positive or negative values
+ - Add Sunrise/Sunset Conditional ([#440](https://github.com/kizniche/mycodo/issues/440))
 
 ### Bugfixes
 
  - Fix issues with PID control on Dashboard ([#441](https://github.com/kizniche/mycodo/issues/441))
+ - Improve LCD controller shutdown speed
 
 
 ## 5.6.10 (2018-03-31)